Fix paths after directory move
