+++ xkcdserver/cmd/xkcd/main.go
@@ -3,10 +3,10 @@ package main
 import (
 	"context"
 	"database/sql"
-	"github.com/AfoninaOlga/xkcd/internal/adapter/client"
-	"github.com/AfoninaOlga/xkcd/internal/adapter/handler"
-	"github.com/AfoninaOlga/xkcd/internal/adapter/repository/sqlite"
-	"github.com/AfoninaOlga/xkcd/internal/core/service"
+	"github.com/AfoninaOlga/xkcd/xkcdserver/internal/adapter/client"
+	"github.com/AfoninaOlga/xkcd/xkcdserver/internal/adapter/handler"
+	"github.com/AfoninaOlga/xkcd/xkcdserver/internal/adapter/repository/sqlite"
+	"github.com/AfoninaOlga/xkcd/xkcdserver/internal/core/service"
 	_ "github.com/golang-migrate/migrate/v4/source/file"
 	_ "github.com/mattn/go-sqlite3"
 	"golang.org/x/sync/errgroup"
