+++ xkcdserver/internal/adapter/client/client_test.go
@@ -3,7 +3,7 @@ package client
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/AfoninaOlga/xkcd/internal/core/domain"
+	"github.com/AfoninaOlga/xkcd/xkcdserver/internal/core/domain"
 	"github.com/stretchr/testify/assert"
 	"net/http"
 	"net/http/httptest"
