+++ xkcdserver/internal/adapter/handler/middleware.go
@@ -2,8 +2,8 @@ package handler
 
 import (
 	"context"
-	"github.com/AfoninaOlga/xkcd/internal/core/domain"
-	"github.com/AfoninaOlga/xkcd/internal/core/port"
+	"github.com/AfoninaOlga/xkcd/xkcdserver/internal/core/domain"
+	"github.com/AfoninaOlga/xkcd/xkcdserver/internal/core/port"
 	"log"
 	"net"
 	"net/http"
