+++ xkcdserver/internal/adapter/handler/xkcd.go
@@ -2,7 +2,7 @@ package handler
 
 import (
 	"encoding/json"
-	"github.com/AfoninaOlga/xkcd/internal/core/port"
+	"github.com/AfoninaOlga/xkcd/xkcdserver/internal/core/port"
 	"log"
 	"net/http"
 	"sync"
