+++ xkcdserver/internal/adapter/handler/xkcd_test.go
@@ -1,7 +1,7 @@
 package handler
 
 import (
-	"github.com/AfoninaOlga/xkcd/internal/core/port/mocks"
+	"github.com/AfoninaOlga/xkcd/xkcdserver/internal/core/port/mocks"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"net/http"
