+++ xkcdserver/internal/core/port/mocks/AuthService.go
@@ -5,7 +5,7 @@ package mocks
 import (
 	context "context"
 
-	domain "github.com/AfoninaOlga/xkcd/internal/core/domain"
+	domain "github.com/AfoninaOlga/xkcd/xkcdserver/internal/core/domain"
 	mock "github.com/stretchr/testify/mock"
 )
